+++ scripts/update_native.sh
@@ -3,6 +3,7 @@ PORT1=4567
 PORT2=5344
 PORT3=5345
 TAG="native"
+UPDATE=false
 MEM_OPT="-Xmx512M"
 MOUNT=""
 
@@ -12,12 +13,13 @@ usage(){
   echo "-p PORT1       管理界面端口，默认：4567"
   echo "-P PORT2       小雅AList端口，默认：5344"
   echo "-t TAG         Docker镜像标签，默认：native"
+  echo "-u             检查镜像更新"
   echo "-v Host:Docker 路径挂载"
   echo "-m MEM_OPT     Java最大堆内存，默认：512M"
   exit 2
 }
 
-while getopts "d:p:P:e:m:t:v:h" arg; do
+while getopts "d:p:P:e:m:t:v:hu" arg; do
     case "${arg}" in
         d)
             BASE_DIR=${OPTARG}
@@ -36,6 +38,9 @@ while getopts "d:p:P:e:m:t:v:h" arg; do
             ;;
         t)
             TAG=${OPTARG}
+            ;;
+        u)
+            UPDATE=true
             ;;
         v)
             MOUNT="${MOUNT} -v ${OPTARG}"
@@ -90,11 +95,18 @@ elif [ "$ARCH" = "aarch64" ]; then
     platform="linux/arm64"
 fi
 
+IMAGE_ID=$(docker images -q haroldli/xiaoya-tvbox:${TAG})
 echo -e "\e[32m下载最新Docker镜像，平台：${platform}, image tag: ${TAG}\e[0m"
 for i in 1 2 3 4 5
 do
    docker pull --platform ${platform} haroldli/xiaoya-tvbox:${TAG} && break
 done
+
+NEW_IMAGE=$(docker images -q haroldli/xiaoya-tvbox:${TAG})
+if [ "$UPDATE" = "true" ] && [ "$IMAGE_ID" = "$NEW_IMAGE" ]; then
+  echo "镜像没有更新"
+  exit
+fi
 
 echo -e "\e[33m重启应用\e[0m"
 docker rm -f xiaoya-tvbox 2>/dev/null && \
